feat: Implement Budgeting CRUD with Generic APIViews and fix some bugs in views

+++ api/views.py
@@ -51,6 +51,8 @@ class CategoryRetrieveUpdateDestroyAPIView(generics.RetrieveUpdateDestroyAPIView
     def preform_destroy(self, instance):
         if instance.transactions.exists():
             raise ValidationError({'detail':'عدم امکان حذف زیرا تراکنش هایی به آن مرنبط است'})
+        if instance.category_budget.exists():
+            raise ValidationError({'detail':'عدم امکان حذف زیرا بودجه بندی هایی برای ان وجود دارند'})
         instance.delete()
 
 class TransactionListCreateAPIView(generics.ListAPIView):
@@ -64,7 +66,7 @@ class TransactionListCreateAPIView(generics.ListAPIView):
         try:
             serializer.save(user=self.request.user)
         except IntegrityError:
-            raise ValidationError({'detail': 'خطای دیتابیس هنگام اپدیت دسته'})
+            raise ValidationError({'detail': 'خطای دیتابیس هنگام ساخت رکورد تراکنش'})
 
 class TransactionRetrieveUpdateDestroyAPIView(generics.RetrieveUpdateDestroyAPIView):
     queryset = Transaction
@@ -77,11 +79,33 @@ class TransactionRetrieveUpdateDestroyAPIView(generics.RetrieveUpdateDestroyAPIV
         try:
             serializer.save(user=self.request.user)
         except IntegrityError:
-            raise ValidationError({'detail': 'خطای دیتابیس هنگام اپدیت دسته'})
+            raise ValidationError({'detail': 'خطای دیتابیس هنگام اپدیت تراکنش ها'})
 
-    def preform_destroy(self, instance):
-        if instance.transactions.exists():
-            raise ValidationError({'detail':'عدم امکان حذف زیرا تراکنش هایی به آن مرنبط است'})
-        instance.delete()
+
+class BudgetingListCreateAPIView(generics.ListCreateAPIView):
+    queryset = Budgeting
+    permission_classes = [IsAuthenticated]
+
+    def get_queryset(self):
+        return Budgeting.objects.select_related('category').filter(user=self.request.user)
+
+    def preform_create(self, serializer):
+        try:
+            serializer.save(user=self.request.user)
+        except IntegrityError:
+            raise ValidationError({'detail': 'خطای دیتابیس هنگام اپدیت بودجه'})
+
+class BudgetingRetrieveUpdateDestroyAPIView(generics.RetrieveUpdateDestroyAPIView):
+    queryset = Budgeting
+    permission_classes = [IsAuthenticated]
+
+    def get_queryset(self):
+        return Budgeting.objects.select_related('category').filter(user=self.request.user)
+
+    def preform_update(self, serializer):
+        try:
+            serializer.save(user=self.request.user)
+        except IntegrityError:
+            raise ValidationError({'detail': 'خطای دیتابیس هنگام اپدیت بودجه'})
